pyproject.toml
[project]
name = "emotion-vectors"
version = "0.1.0"
description = "Extract and validate emotion vectors from Gemma 4's internal representations"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.48.0",
    "accelerate>=1.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "jsonlines>=4.0.0",
    "umap-learn>=0.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
